fix(go): remove old references.
