+++ examples/server/handlers/dcr.go
@@ -33,7 +33,7 @@ import (
 	corev1 "zntr.io/solid/api/gen/go/oidc/core/v1"
 	"zntr.io/solid/api/oidc"
 	"zntr.io/solid/pkg/authorizationserver"
-	"zntr.io/solid/pkg/rfcerrors"
+	"zntr.io/solid/pkg/sdk/rfcerrors"
 )
 
 // DCR handles dynamic client registration.
