+++ go.mod
@@ -7,14 +7,13 @@ require (
 	github.com/fatih/color v1.9.0
 	github.com/golang/mock v1.4.3
 	github.com/golang/protobuf v1.4.2
-	github.com/google/go-cmp v0.4.1
-	github.com/imdario/mergo v0.3.9
+	github.com/google/go-cmp v0.5.1
 	github.com/kr/session v0.1.0
-	github.com/magefile/mage v1.9.0
+	github.com/magefile/mage v1.10.0
 	github.com/patrickmn/go-cache v2.1.0+incompatible
 	github.com/spf13/cobra v1.0.0
-	github.com/square/go-jose/v3 v3.0.0-20200603004136-8ccb8a19e809
-	golang.org/x/crypto v0.0.0-20200604202706-70a84ac30bf9
+	github.com/square/go-jose/v3 v3.0.0-20200630053402-0a67ce9b0693
+	golang.org/x/crypto v0.0.0-20200709230013-948cd5f35899
 	golang.org/x/oauth2 v0.0.0-20200107190931-bf48bf16ab8d
-	google.golang.org/protobuf v1.24.0
+	google.golang.org/protobuf v1.25.0
 )
