+++ internal/services/client/service.go
@@ -31,10 +31,10 @@ import (
 	corev1 "zntr.io/solid/api/gen/go/oidc/core/v1"
 	"zntr.io/solid/api/oidc"
 	"zntr.io/solid/internal/services"
-	"zntr.io/solid/pkg/profile"
-	"zntr.io/solid/pkg/rfcerrors"
-	"zntr.io/solid/pkg/storage"
-	"zntr.io/solid/pkg/types"
+	"zntr.io/solid/pkg/sdk/rfcerrors"
+	"zntr.io/solid/pkg/sdk/types"
+	"zntr.io/solid/pkg/server/profile"
+	"zntr.io/solid/pkg/server/storage"
 )
 
 type service struct {
